requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==4.1.1
djangorestframework==3.13.1
drf-spectacular==0.24.0
inflection==0.5.1
jsonschema==4.16.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pydantic==1.10.2
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
